app/pyproject.toml
[tool.poetry]
name = "v2-interview"
version = "0.1.0"
description = ""
authors = ["nazarkobryn <kobrinnazar05@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^5.1.3"
django-webpack-loader = "^3.1.1"
djangorestframework = "^3.15.2"
django-extensions = "^3.2.3"
drf-yasg = "^1.21.8"


[tool.poetry.group.dev.dependencies]
ruff = "^0.7.4"
pytest = "^8.3.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
